refactor: extract ToggleBaseCover from ToggleModeCover
